diamond_price_predictor_ml/pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
# This is the project/package name for PyPI and Poetry
name = "diamond_price_predictor_ml"
version = "0.0.1"
description = "Diamond price predictor ML package"
authors = [
  { name = "Your Name" }
]
readme = "README.md"
requires-python = ">=3.11,<4.0"

# This tells Poetry which folder is the actual Python package
packages = [
    { include = "diamond_price_predictor_ml" }
]

# Dependencies
dependencies = [
    "ipython",
    "jupyterlab",
    "loguru",
    "matplotlib",
    "mkdocs",
    "notebook",
    "numpy",
    "pandas",
    "pip",
    "python-dotenv",
    "ruff",
    "scikit-learn",
    "tqdm",
    "typer",
    "mlflow>=3.3.2,<4.0.0",
    "seaborn>=0.13.2,<0.14.0",
    "pydantic-settings>=2.10.1,<3.0.0",
    "statsmodels>=0.14.5,<0.15.0",
    "databricks-sdk>=0.67.0,<0.68.0"
]

classifiers = [
    "Programming Language :: Python :: 3",
]

[tool.ruff]
line-length = 99
src = ["ml"]
include = ["pyproject.toml", "ml/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["ml"]
force-sort-within-sections = true
